# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Agents - HTTP API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.27.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/ensemble/chat": {
      "post": {
        "tags": [
          "ensemble"
        ],
        "summary": "Improv turn request\nCreate chat session",
        "operationId": "create_chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Chat session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnsembleResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/ensemble/chat/{session}/improv": {
      "post": {
        "tags": [
          "ensemble"
        ],
        "summary": "Execute an improvisation turn in a chat session",
        "operationId": "improv_turn",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImprovTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Improv turn completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImprovTurnResponse"
                }
              }
            }
          },
          "404": {
            "description": "Chat session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/ensemble/standing-start": {
      "post": {
        "tags": [
          "ensemble"
        ],
        "summary": "Start a standing ensemble session with initial context",
        "operationId": "standing_start",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StandingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Standing session started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingStartResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/ensemble/standing-status": {
      "get": {
        "tags": [
          "ensemble"
        ],
        "summary": "Get standing session status",
        "operationId": "standing_status",
        "responses": {
          "200": {
            "description": "Standing session status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandingStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No standing session found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateChatRequest": {
        "type": "object",
        "description": "Create chat request",
        "required": [
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "EnsembleResponse": {
        "type": "object",
        "description": "Chat response",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ImprovTurnRequest": {
        "type": "object",
        "description": "Improv turn request",
        "required": [
          "user_message"
        ],
        "properties": {
          "user_message": {
            "type": "string"
          }
        }
      },
      "ImprovTurnResponse": {
        "type": "object",
        "description": "Improv turn response",
        "required": [
          "user_message",
          "judgment_count",
          "response_count"
        ],
        "properties": {
          "curator_synthesis": {
            "type": [
              "string",
              "null"
            ]
          },
          "judgment_count": {
            "type": "integer",
            "minimum": 0
          },
          "response_count": {
            "type": "integer",
            "minimum": 0
          },
          "user_message": {
            "type": "string"
          }
        }
      },
...
      "ParticipantEntryRequest": {
        "type": "object",
        "description": "Participant entry for standing session start",
        "required": [
          "agent",
          "type",
          "role",
          "description"
        ],
        "properties": {
          "agent": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Template domains this participant owns (e.g., `[\"cns\", \"storage\"]`).\nConverted to capability specs (`\"tool:<domain>:execute\"`) internally."
          },
          "role": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "ParticipantStatusResponse": {
        "type": "object",
        "description": "Participant status in standing session response",
        "required": [
          "name",
          "webid",
          "role",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "webid": {
            "type": "string"
          }
        }
      },
...
      "StandingStartRequest": {
        "type": "object",
        "description": "Standing session start request",
        "required": [
          "session_id",
          "name",
          "description",
          "initial_context",
          "participants"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "initial_context": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantEntryRequest"
            }
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "StandingStartResponse": {
        "type": "object",
        "description": "Standing session start response",
        "required": [
          "session_id",
          "description",
          "participant_count",
          "message"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "participant_count": {
            "type": "integer",
            "minimum": 0
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "StandingStatusResponse": {
        "type": "object",
        "description": "Standing session status response",
        "required": [
          "session_id",
          "description",
          "participant_count",
          "message_count",
          "participants"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "message_count": {
            "type": "integer",
            "minimum": 0
          },
          "participant_count": {
            "type": "integer",
            "minimum": 0
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantStatusResponse"
            }
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}